Add new feedback collection items
