+++ src/Razor/test/Microsoft.VisualStudio.Razor.IntegrationTests/BreakpointSpanTests.cs
@@ -12,17 +12,17 @@ namespace Microsoft.VisualStudio.Razor.IntegrationTests
         public async Task SetBreakpoint_FirstCharacter_SpanAdjusts()
         {
             // Open the file
-            await TestServices.SolutionExplorer.OpenFileAsync(BlazorProjectName, CounterRazorFile, HangMitigatingCancellationToken);
+            await TestServices.SolutionExplorer.OpenFileAsync(RazorProjectConstants.BlazorProjectName, RazorProjectConstants.CounterRazorFile, HangMitigatingCancellationToken);
 
             // Wait for classifications to indicate Razor LSP is up and running
             await TestServices.Editor.WaitForClassificationAsync(HangMitigatingCancellationToken);
             await TestServices.Editor.SetTextAsync("<p>@{ var abc = 123; }</p>", HangMitigatingCancellationToken);
 
             // Act
-            await TestServices.Debugger.SetBreakpointAsync(CounterRazorFile, line: 1, character: 1, HangMitigatingCancellationToken);
+            await TestServices.Debugger.SetBreakpointAsync(RazorProjectConstants.CounterRazorFile, line: 1, character: 1, HangMitigatingCancellationToken);
 
             // Assert
-            await TestServices.Debugger.VerifyBreakpointAsync(CounterRazorFile, line: 1, character: 7, HangMitigatingCancellationToken);
+            await TestServices.Debugger.VerifyBreakpointAsync(RazorProjectConstants.CounterRazorFile, line: 1, character: 7, HangMitigatingCancellationToken);
         }
 
         [IdeFact]
@@ -39,7 +39,7 @@ namespace Microsoft.VisualStudio.Razor.IntegrationTests
             }
 
             // Open the file
-            await TestServices.SolutionExplorer.OpenFileAsync(BlazorProjectName, CounterRazorFile, HangMitigatingCancellationToken);
+            await TestServices.SolutionExplorer.OpenFileAsync(RazorProjectConstants.BlazorProjectName, RazorProjectConstants.CounterRazorFile, HangMitigatingCancellationToken);
 
             // Wait for classifications to indicate Razor LSP is up and running
             await TestServices.Editor.WaitForClassificationAsync(HangMitigatingCancellationToken);
@@ -48,7 +48,7 @@ namespace Microsoft.VisualStudio.Razor.IntegrationTests
 }</p>", HangMitigatingCancellationToken);
 
             // Act
-            var result = await TestServices.Debugger.SetBreakpointAsync(CounterRazorFile, line: 1, character: 1, HangMitigatingCancellationToken);
+            var result = await TestServices.Debugger.SetBreakpointAsync(RazorProjectConstants.CounterRazorFile, line: 1, character: 1, HangMitigatingCancellationToken);
 
             // Assert
             Assert.False(result);
@@ -58,7 +58,7 @@ namespace Microsoft.VisualStudio.Razor.IntegrationTests
         public async Task SetBreakpoint_FirstCharacter_ValidLine()
         {
             // Open the file
-            await TestServices.SolutionExplorer.OpenFileAsync(BlazorProjectName, CounterRazorFile, HangMitigatingCancellationToken);
+            await TestServices.SolutionExplorer.OpenFileAsync(RazorProjectConstants.BlazorProjectName, RazorProjectConstants.CounterRazorFile, HangMitigatingCancellationToken);
 
             // Wait for classifications to indicate Razor LSP is up and running
             await TestServices.Editor.WaitForClassificationAsync(HangMitigatingCancellationToken);
@@ -67,10 +67,10 @@ namespace Microsoft.VisualStudio.Razor.IntegrationTests
 }</p>", HangMitigatingCancellationToken);
 
             // Act
-            await TestServices.Debugger.SetBreakpointAsync(CounterRazorFile, line: 2, character: 1, HangMitigatingCancellationToken);
+            await TestServices.Debugger.SetBreakpointAsync(RazorProjectConstants.CounterRazorFile, line: 2, character: 1, HangMitigatingCancellationToken);
 
             // Assert
-            await TestServices.Debugger.VerifyBreakpointAsync(CounterRazorFile, line: 2, character: 4, HangMitigatingCancellationToken);
+            await TestServices.Debugger.VerifyBreakpointAsync(RazorProjectConstants.CounterRazorFile, line: 2, character: 4, HangMitigatingCancellationToken);
         }
     }
 }
